requirements.txt
streamlit>=1.35.0
openai>=1.30.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
python-pptx>=0.6.23
